+++ R/readClusterDesc.R
@@ -37,10 +37,21 @@
 #' \code{readClusterSTDesc} : read st-storage clusters (Antares >= V8.6)
 #' 
 #' If you have no clusters properties, `Null data.table (0 rows and 0 cols)` is returned.
+#' 
+#' @section Warning:  
+#' You have now two format output to display input properties. 
+#' Default is format uses by operating team, eg `min.down.time`.
+#' Other format is according to antares simulator, eg `min-down-time`.  
+#' 
+#' All properties are returned with default values according to Antares Study version.
+#' 
+#' 
 #'
 #' @examples
 #' 
 #' \dontrun{
+#' 
+#' # Default format with "dot separator"
 #' 
 #' # thermal
 #' readClusterDesc()
@@ -50,6 +61,18 @@
 #' 
 #' # st-storage
 #' readClusterSTDesc()
+#' 
+#' # Antares Simulator format 
+#' 
+#' #' # thermal
+#' readClusterDesc(dot_format = FALSE)
+#' 
+#' # renewable
+#' readClusterResDesc(dot_format = FALSE)
+#' 
+#' # st-storage
+#' readClusterSTDesc(dot_format = FALSE)
+#' 
 #' 
 #' # By default, the function reads cluster descriptions for the default study,
 #' # but it is possible to specify another study with parameter "opts"