+++ NEWS.md
@@ -6,6 +6,7 @@
 
 BUGFIXES :
 * `setSimulationPathAPI()` : encode URL before reading the data in simulation mode
+* `readAntares()` : In disk mode, return all the available columns for a short-term storage output and match the column with the content
 
 BREAKING CHANGES :  
 * `setSimulationPathAPI()` : reads and returns the new converted study version format (ex : 9.0 => 900)