+++ NEWS.md
@@ -1,7 +1,7 @@
 > Copyright © 2016 RTE Réseau de transport d’électricité
 
 
-# antaresRead 2.8.0.9000 
+# antaresRead 2.9.0
 (cf. Antares v9 changelog)  
 
 NEW FEATURES: